@@ -10,8 +10,12 @@ class Development(commands.GroupCog, name="dev"):
         super().__init__()
 
     async def auto_complete_cog(self, inter: Interaction, current: str):
-        cog_files = glob.glob('rank_keeper/cogs/*.py')
-        return [app_commands.Choice(name=path, value=path) for path in cog_files if path.lower() in current.lower()]
+        cog_files = glob.glob("rank_keeper/cogs/*.py")
+        return [
+            app_commands.Choice(name=path, value=path)
+            for path in cog_files
+            if current.lower() in path.lower()
+        ]
 
     @app_commands.command(name="load", description="Cogをロードします。")
     @app_commands.autocomplete(extension=auto_complete_cog)
@@ -19,7 +23,9 @@ class Development(commands.GroupCog, name="dev"):
         try:
             await self.bot.load_extension(extension)
         except commands.ExtensionAlreadyLoaded:
-            return await inter.response.send_message(f'{extension}は既に読み込まれています。')
+            return await inter.response.send_message(
+                f"{extension}は既に読み込まれています。"
+            )
         await self.bot.tree.sync()
         await inter.response.send_message(
             f"{extension} loaded.", ephemeral=True
@@ -31,7 +37,9 @@ class Development(commands.GroupCog, name="dev"):
         try:
             await self.bot.unload_extension(extension)
         except commands.ExtensionNotLoaded:
-            return await inter.response.send_message(f'{extension}は読み込まれていません。')
+            return await inter.response.send_message(
+                f"{extension}は読み込まれていません。"
+            )
         await self.bot.tree.sync()
         await inter.response.send_message(
             f"{extension} unloaded.", ephemeral=True
@@ -43,7 +51,9 @@ class Development(commands.GroupCog, name="dev"):
         try:
             await self.bot.reload_extension(extension)
         except commands.ExtensionNotLoaded:
-            return await inter.response.send_message(f'{extension}は読み込まれていません。')
+            return await inter.response.send_message(
+                f"{extension}は読み込まれていません。"
+            )
         await self.bot.tree.sync()
         await inter.response.send_message(
             f"{extension} reloaded.", ephemeral=True
